Refactor calculateCredit and related tests
